+++ client/src/pages/orders/index.vue
@@ -1,4 +1,11 @@
 <template>
+  <!-- Alert -->
+  <AlertComponent
+    :alert-active="showDeleteAlert"
+    :entity="'order(s)'"
+    :on-confirm="deleteOrder"
+    @close-alert="closeAlert"
+  />
   <!-- Modal -->
   <OrderModal
     :updatedOrder="updatedOrder"
@@ -10,7 +17,8 @@
   <button
     @click="openModal()"
     class="w-full md:w-2/5 bg-white border-b border-gray-300 hover:border-gray-700 hover:bg-gray-800 hover:text-white dark:border-gray-700 dark:bg-gray-800 dark:text-white dark:hover:bg-white dark:hover:border-gray-300 dark:hover:text-gray-800 font-bold py-2 px-4 mb-4 rounded"
-  ><font-awesome-icon icon="fa-solid fa-cart-plus" />
+  >
+    <font-awesome-icon icon="fa-solid fa-cart-plus" />
     Add new order
   </button>
 
@@ -30,7 +38,7 @@
     :selectedCount="selectedCount"
     :export-selected="exportLink"
     @update-item="openModal"
-    @delete-item="deleteOrder"
+    @delete-item="confirmDelete"
   >
     <template #mobile-sort-menu>
       <div class="w-full text-left px-4 py-2 text-gray-800">
@@ -98,7 +106,7 @@
           Update
         </button>
         <button
-          @click="deleteCustomer(item)"
+          @click="confirmDelete(item)"
           class="text-red-600 hover:text-red-800 hover:font-medium cursor-pointer"
         >
           Delete
@@ -189,7 +197,7 @@
             Update
           </button>
           <button
-            @click="deleteOrder(item)"
+            @click="confirmDelete(item)"
             class="text-red-400 hover:text-red-600 hover:font-medium cursor-pointer ml-1"
           >
             Delete
@@ -221,6 +229,7 @@
 </template>
 
 <script>
+import AlertComponent from "@/components/AlertComponet.vue";
 import { useOrderStore } from "@/store/order.js";
 import OrderModal from "./components/OrderModal.vue";
 import SearchComponent from "@/components/SearchComponent.vue";
@@ -233,9 +242,11 @@ export default {
     SearchComponent,
     TableComponent,
     SortIconsComponent,
+    AlertComponent,
   },
   data() {
     return {
+      showDeleteAlert: false,
       orders: [],
       updatedOrder: {},
       openOrderModal: false,
@@ -253,6 +264,7 @@ export default {
       },
       sortHeader: "",
       searchActive: false,
+      itemsToDelete: null,
     };
   },
   created() {
@@ -324,13 +336,25 @@ export default {
     closeModal() {
       this.openOrderModal = false;
     },
-    async deleteOrder(order) {
-      if (this.selectedItems.length > 0) {
-        await this.orderStore.deleteOrder(this.selectedItems.join(","));
-        this.selectedItems.length = 0;
-      } else {
-        await this.orderStore.deleteOrder(order.id);
+    closeAlert() {
+      this.showDeleteAlert = false;
+    },
+    confirmDelete(item) {
+      this.itemsToDelete = item;
+      this.showDeleteAlert = true;
+    },
+    async deleteOrder() {
+      if (this.itemsToDelete) {
+        if (this.selectedItems.length > 0) {
+          await this.orderStore.deleteOrder(this.selectedItems.join(","));
+          this.selectedItems.length = 0;
+        } else {
+          await this.orderStore.deleteOrder(this.itemsToDelete.id);
+        }
+        this.showDeleteAlert = false;
+        this.itemsToDelete = null;
       }
+
       this.loadOrders();
     },
     async sortColumn(column) {
