+++ client/src/pages/products/index.vue
@@ -1,4 +1,12 @@
 <template>
+  <!-- Alert -->
+  <AlertComponent
+    :alert-active="showDeleteAlert"
+    :entity="'product(s)'"
+    :on-confirm="deleteProduct"
+    @close-alert="closeAlert"
+  />
+  <!-- Modal -->
   <ProductModal
     :updatedProduct="updatedProduct"
     :modalActive="openProductModal"
@@ -11,7 +19,8 @@
   <button
     @click="openModal()"
     class="w-full md:w-2/5 bg-white border-b border-gray-300 hover:border-gray-700 hover:bg-gray-800 hover:text-white dark:border-gray-700 dark:bg-gray-800 dark:text-white dark:hover:bg-white dark:hover:border-gray-300 dark:hover:text-gray-800 font-bold py-2 px-4 mb-4 rounded"
-  ><font-awesome-icon icon="fa-solid fa-plus" />
+  >
+    <font-awesome-icon icon="fa-solid fa-plus" />
     Add new product
   </button>
 
@@ -31,7 +40,7 @@
     :selectedCount="selectedCount"
     :export-selected="exportLink"
     @update-item="openModal"
-    @delete-item="deleteProduct"
+    @delete-item="confirmDelete"
   >
     <!-- Mobile cards -->
     <template #mobile-sort-menu>
@@ -100,7 +109,7 @@
           Update
         </button>
         <button
-          @click="deleteProduct(item)"
+          @click="confirmDelete(item)"
           class="text-red-600 hover:text-red-800 hover:font-medium cursor-pointer"
         >
           Delete
@@ -192,7 +201,7 @@
             Update
           </button>
           <button
-            @click="deleteProduct(item)"
+            @click="confirmDelete(item)"
             class="text-red-400 hover:text-red-600 hover:font-medium cursor-pointer ml-1"
           >
             Delete
@@ -225,6 +234,7 @@
 </template>
 
 <script>
+import AlertComponent from "@/components/AlertComponet.vue";
 import { useProductStore } from "@/store/product.js";
 import ProductModal from "./components/ProductModal.vue";
 import SearchComponent from "@/components/SearchComponent.vue";
@@ -237,9 +247,11 @@ export default {
     SearchComponent,
     TableComponent,
     SortIconsComponent,
+    AlertComponent,
   },
   data() {
     return {
+      showDeleteAlert: false,
       products: [],
       pageSize: 10,
       displayedProducts: [],
@@ -255,6 +267,7 @@ export default {
       },
       sortHeader: "",
       searchActive: false,
+      itemsToDelete: null,
     };
   },
   created() {
@@ -300,13 +313,25 @@ export default {
       ];
       this.loadedProductsCount += remainingProducts.length;
     },
-    async deleteProduct(product) {
-      if (this.selectedItems.length > 0) {
-        await this.productStore.deleteProduct(this.selectedItems);
-        this.selectedItems.length = 0;
-      } else {
-        await this.productStore.deleteProduct(product.id);
+    closeAlert() {
+      this.showDeleteAlert = false;
+    },
+    confirmDelete(item) {
+      this.itemsToDelete = item;
+      this.showDeleteAlert = true;
+    },
+    async deleteProduct() {
+      if (this.itemsToDelete) {
+        if (this.selectedItems.length > 0) {
+          await this.productStore.deleteProduct(this.selectedItems);
+          this.selectedItems.length = 0;
+        } else {
+          await this.productStore.deleteProduct(this.itemsToDelete.id);
+        }
+        this.showDeleteAlert = false;
+        this.itemsToDelete = null;
       }
+
       this.loadProducts();
     },
     async updateDisplayedProducts(searchFilter) {
